requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
Markdown==3.3.2
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.4.1
toml==0.10.1
mysqlclient==2.0.1
requests==2.24.0